requirements.txt
pandas
pyautogui
pyperclip
pyinstaller
python-dotenv
openpyxl
tk
pillow
cairosvg
svgwrite
pytest
python-dotenv